Actual first map, main menu, outlines and decorations, egg cannon model

+++ scripts/egg_modules/breaker.gd
@@ -10,12 +10,17 @@ onready var powerups = get_node("/root/Main/Powerups")
 
 onready var status = get_node("../Status")
 
+var epsilon = 0.05
+
 func _integrate_forces(state):
 	if status.delivered: return
 	
 	for i in range(state.get_contact_count()):
 		var obj = state.get_contact_collider_object(i)
 		if not obj.is_in_group("Breakers"): continue
+		
+		var normal = state.get_contact_local_normal(i)
+		if normal.y < (0 + epsilon): return
 		
 		destroy_myself()
 		break
